requirements.txt
streamlit
conch @ git+https://github.com/owner/repo@main
pandas 
huggingface_hub
torch
Pillow